requirements.txt
xmltodict==0.13.0
pandas==2.0.1
requests==2.31.0
apache-airflow-providers-google==10.16.0